Bug fix: update in-memory high_ after saving score and bump version to 1.0.1

+++ game.cpp
@@ -281,7 +281,7 @@ void Game::drawBoxNoFill() const {
 }
 
 void Game::gameOver() {
-  score_.saveScore((int)score_.current_);
+  score_.saveScore(score_.current_);
   gameOverDraw();
 }
